feat(automodule): generate in-between options
When transitioning from option a.b.c to a.d.e.f, it generates mock
options a.d and a.d.e

Also generate a mock option for the top-level module passed to
automodule.

This enables referring to modules containing options, but are not
themselves options, and this fixes the issue where in the TOC the
hierarchy was not reported when a level is skipped (fixes #7).